requirements.txt
pytest
boto3
pyjwt
requests
botocore
